[scripts] Avoid splitting on &nbsp (appears in some UTF-8 chars) when tokenizing lexicon (#2676)
Some UTF-8 characters (for example Š) are interpreted in latin-1 as containing whitespace. For example, Š consists of bytes c5 and a0 in UTF-8, but a0 corresponds to non-breaking space in latin-1. This means that words containing Š are not tokenized correctly when reading lexicon.txt.

This change ensures that lexicon.txt is tokenized using only space or tab characters.